+++ resources/views/manager/position.blade.php
@@ -42,7 +42,21 @@
                             Edit
                         </button>
 
-                        <a href="/position/delete/{{$position->id}}" class="btn btn-danger">Delete</a>
+                        <a href="/position/delete/{{$position->id}}" class="btn btn-danger" onclick="confirmDelete(event)">Delete</a>
+                        {{--Script Delete--}}
+                        <script>
+                            function confirmDelete(event) {
+                                event.preventDefault();
+
+                                const result = confirm('Are you sure you want to delete this position?');
+
+                                if (result === true) {
+                                    window.location.href = event.target.href;
+                                } else {
+                                    alert('Delete canceled!');
+                                }
+                            }
+                        </script>
                     </td>
                 </tr>
                 {{--    <!-- Modal -->--}}
@@ -64,7 +78,21 @@
                                 </div>
                                 <div class="modal-footer">
                                     <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Đóng</button>
-                                    <button type="submit" class="btn btn-primary">thay doi</button>
+                                    <button type="submit" id="editButton_{{ $position->id }}" class="btn btn-primary" onclick="confirmAndUpdate_{{ $position->id }}(event)">Thay Đổi</button>
+                                    {{--Script edit--}}
+                                    <script>
+                                        function confirmAndUpdate_{{ $position->id }}(event) {
+                                            event.preventDefault();
+
+                                            const result = confirm('Are you sure you want to position this level ?');
+
+                                            if (result === true) {
+                                                event.target.form.submit();
+                                            } else {
+                                                alert('Update position canceled!');
+                                            }
+                                        }
+                                    </script>
                                 </div>
                             </div>
                         </form>
@@ -85,11 +113,26 @@
                             </div>
                             <div class="modal-body">
                                 Name level<input class="form-control" type="text" value="" name="position_name" required>
-                                Salary Factor<input class="form-control mt-2" type="number" name="salary" min="1.0" step="0.1" required>
+                                Salary <input class="form-control mt-2" type="number" name="salary" min="1.0" step="0.1" required>
                             </div>
                             <div class="modal-footer">
                                 <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Đóng</button>
-                                <button type="submit" class="btn btn-primary">Tạo mới</button>
+                                <button type="submit" id="addButton" class="btn btn-primary" onclick="addPositionData(event)">Tạo mới</button>
+                                {{--Script add--}}
+                                <script>
+                                    function addPositionData(event) {
+                                        event.preventDefault();
+
+                                        // Hiển thị cảnh báo
+                                        const result = confirm('Bạn có chắc chắn muốn thêm dữ liệu Position mới?');
+                                        if (result === true) {
+                                            // Gửi biểu mẫu
+                                            event.target.form.submit();
+                                        } else {
+                                            alert('Thêm dữ liệu mới đã bị hủy!');
+                                        }
+                                    }
+                                </script>
                             </div>
                         </div>
                     </form>
